+++ mapchete/log.py
@@ -9,6 +9,26 @@ correctly.
 import logging
 
 
+# lower stream output log level
+formatter = logging.Formatter('%(asctime)s %(levelname)s %(name)s %(message)s')
+stream_handler = logging.StreamHandler()
+stream_handler.setFormatter(formatter)
+stream_handler.setLevel(logging.ERROR)
+logging.getLogger().addHandler(stream_handler)
+
+
+def set_log_level(loglevel):
+    logging.getLogger("mapchete").setLevel(loglevel)
+    stream_handler.setLevel(loglevel)
+
+
+def setup_logfile(logfile):
+    file_handler = logging.FileHandler(logfile)
+    file_handler.setFormatter(formatter)
+    logging.getLogger().addHandler(file_handler)
+    logging.getLogger("mapchete").setLevel(logging.DEBUG)
+
+
 def user_process_logger(pname):
     """Logger to be used within a user process file."""
     return logging.getLogger("mapchete.user_process." + pname)